bgp-analyzer/pyproject.toml
[project]
name = "netvantage-bgp-analyzer"
version = "0.1.0"
description = "BGP Analysis Engine for NetVantage — hijack detection, RPKI validation, routing anomaly alerting"
requires-python = ">=3.10"
license = {text = "BSL-1.1"}

dependencies = [
    "prometheus-client>=0.20.0",
    "requests>=2.31.0",
    "structlog>=24.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# pybgpstream requires libparsebgp C library — install system deps first.
# See bgp-analyzer/Dockerfile for the full dependency list.
stream = [
    "pybgpstream>=3.0.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
